Add play trailer component
