+++ src/components/banner/Banner.css
@@ -61,3 +61,7 @@
   padding-top: 20px;
   padding-left: 20px;
 }
+
+.banner__button {
+font-weight: 600 ;
+}
